smart_mail/requirements.txt
tqdm
elasticsearch
boto3
boto3-stubs[bedrock-runtime]
streamlit
requests
sentence-transformers
psycopg2-binary
types-psycopg2
types-requests
sqlalchemy
python-dotenv
pandas-stubs